Start refactoring pipeline scripts
