+++ pipeline/create_training_set.R
