+++ pipeline/requirements.txt
@@ -1,5 +1,6 @@
 argcomplete==1.12.3
-argon2-cffi==21.1.0
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
 attrs==21.2.0
 backcall==0.2.0
 bleach==4.1.0
@@ -17,16 +18,16 @@ fsspec==2021.11.1
 graphviz==0.19.1
 greenlet==1.1.2
 HeapDict==1.0.1
-importlib-metadata==4.8.2
+importlib-metadata==4.9.0
 importlib-resources==5.4.0
 ipykernel==6.6.0
-ipython==7.30.0
+ipython==7.30.1
 ipython-genutils==0.2.0
 ipython-sql==0.4.0
 jedi==0.18.1
 Jinja2==3.0.3
 joblib==1.1.0
-jsonschema==4.2.1
+jsonschema==4.3.0
 jupyter-client==7.1.0
 jupyter-core==4.9.1
 jupyterlab-pygments==0.1.2
@@ -42,19 +43,17 @@ nest-asyncio==1.5.4
 notebook==6.4.6
 numpy==1.21.4
 packaging==21.3
-pandas==1.3.4
+pandas==1.3.5
 pandocfilters==1.5.0
 parso==0.8.3
 partd==1.2.0
 pexpect==4.8.0
-pgspecial==1.13.0
 pickleshare==0.7.5
 Pillow==8.4.0
 prettytable==0.7.2
 prometheus-client==0.12.0
-prompt-toolkit==3.0.23
+prompt-toolkit==3.0.24
 psutil==5.8.0
-psycopg2==2.9.2
 psycopg2-binary==2.9.2
 ptyprocess==0.7.0
 pycparser==2.21
@@ -71,7 +70,7 @@ Send2Trash==1.8.0
 six==1.16.0
 sklearn==0.0
 sortedcontainers==2.4.0
-SQLAlchemy==1.4.27
+SQLAlchemy==1.4.28
 sqlparse==0.4.2
 tblib==1.7.0
 terminado==0.12.1
